requirements.txt
tensorflow-gpu===1.8.0
foolbox==1.1.0
numpy
git+https://github.com/bveliqi/adversarial-vision-challenge
